backend/requirements.txt
fastapi
uvicorn
motor
pydantic
pydantic-settings
pyjwt
passlib[argon2]
python-multipart